Modify Numba JIT calls in _gp.py
In order to get access to scipy's linalg.cho_solve, we needed to determine which functions were necessary to JIT. Since most of the computational work was done in computing the kernel and jittering the matrix, we were able to remove the @njit calls for most of the _detrend_all_signals_gp_numba stack. This resulted in better performance, particularly for larger signals, and perhaps better numerical stability.